Fix typing for the middleware
